+++ proofsharing/__init__.py
@@ -2,4 +2,10 @@ import sys
 import os
 sys.path = [os.path.dirname(__file__)] + sys.path
 
-from utils import *
+from .check_proof_subsumption import *
+from .config import *
+from .utils import *
+from .templates import *
+from .relaxations import *
+from .networks import *
+from .models import *
