--- proofsharing/utils/__init__.py
@@ -1,7 +1,0 @@
-from .utils import *
-from .templates import *
-from .relaxations import *
-from .networks import *
-from .models import *
-from .config import *
-from check_proof_subsumption import *
